Solve small bugs
