+++ src/services/azureService.js
@@ -1,114 +1,19 @@
-import qs from 'querystring';
-import moment from 'moment';
+import AuthService from './authService';
 import api from './api';
-import secretConfig from '../config.secret.json';
 
-export default class AzureService {
-  static callbackUrl = `${secretConfig.url}/callback.html`;
-
+export default class AzureService extends AuthService {
   constructor(store, organization, project, team) {
+    super(store, organization, project, team);
+
     this.organization = organization;
     this.project = project;
     this.team = team;
     this.apiVersion = '5.0';
-    this.code = '';
-    this.accessToken = '';
-    this.refreshToken = '';
-    this.expiresIn = moment();
-    this.onAuthorized = () => store.dispatch('authorizeApp');
-  }
-
-  async initData() {
-    const accessToken = window.localStorage.getItem('accessToken');
-    if (accessToken) {
-      this.accessToken = accessToken;
-      this.refreshToken = window.localStorage.getItem('refreshToken');
-      this.expiresIn = moment(window.localStorage.getItem('expiresIn'));
-
-      if (this.expiresIn.isBefore(moment())) {
-        this.refreshAccessToken();
-      }
-      return;
-    }
-
-    const callbackData = window.localStorage.getItem('callbackData');
-    if (callbackData) {
-      window.localStorage.removeItem('callbackData');
-      await this.handleCallback(callbackData);
-    }
-  }
-
-  static get authUrl() {
-    const params = {
-      client_id: secretConfig.clientId,
-      response_type: 'Assertion',
-      state: 'none',
-      scope: 'vso.code vso.project vso.work',
-      redirect_uri: AzureService.callbackUrl,
-    };
-    return `https://app.vssps.visualstudio.com/oauth2/authorize?${qs.stringify(params)}`;
-  }
-
-  get authHeader() {
-    return { Authorization: `Bearer ${this.accessToken}` };
-  }
-
-  async refreshAccessToken() {
-    const result = await api.post({
-      url: 'https://app.vssps.visualstudio.com/oauth2/token',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded'
-      },
-      data: qs.stringify({
-        client_assertion_type: 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-        client_assertion: secretConfig.appSecret,
-        grant_type: 'refresh_token',
-        assertion: this.refreshToken,
-        redirect_uri: AzureService.callbackUrl
-      })
-    });
-
-    if (!result) {
-      return;
-    }
-
-    this.saveTokenResponse(result);
-  }
-
-  async handleCallback(data) {
-    const urlParams = new URLSearchParams(data);
-    if (urlParams.has('code')) {
-      this.code = urlParams.get('code');
-      this.onAuthorized();
-    }
-
-    await this.getAccessToken();
-  }
-
-  async getAccessToken() {
-    const result = await api.post({
-      url: 'https://app.vssps.visualstudio.com/oauth2/token',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded'
-      },
-      data: qs.stringify({
-        client_assertion_type: 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-        client_assertion: secretConfig.appSecret,
-        grant_type: 'urn:ietf:params:oauth:grant-type:jwt-bearer',
-        assertion: this.code,
-        redirect_uri: AzureService.callbackUrl
-      })
-    });
-
-    if (!result) {
-      return;
-    }
-
-    this.code = '';
-    this.saveTokenResponse(result);
   }
 
   async getIterationWorkItems(iterationId) {
+    await this.checkTokenExpiration();
+
     const result = await api.get({
       url: `https://dev.azure.com/${this.organization}/${this.project}/${this.team}/_apis/work/teamsettings/iterations/${iterationId}/workitems`,
       params: {
@@ -125,6 +30,8 @@ export default class AzureService {
   }
 
   async getWorkItemDetails(itemReferences) {
+    await this.checkTokenExpiration();
+
     const pbiIds = itemReferences
       .filter(x => !x.rel)
       .map(x => x.target.id)
@@ -174,6 +81,8 @@ export default class AzureService {
   }
 
   async getCurrentIteration() {
+    await this.checkTokenExpiration();
+
     const result = await api.get({
       url: `https://dev.azure.com/${this.organization}/${this.project}/${this.team}/_apis/work/teamsettings/iterations`,
       params: {
@@ -188,15 +97,5 @@ export default class AzureService {
     }
 
     return result.data.value[0];
-  }
-
-  saveTokenResponse(result) {
-    this.accessToken = result.data.access_token;
-    this.refreshToken = result.data.refresh_token;
-    this.expiresIn = moment().add(result.data.expires_in, 'seconds');
-
-    window.localStorage.setItem('accessToken', this.accessToken);
-    window.localStorage.setItem('refreshToken', this.refreshToken);
-    window.localStorage.setItem('expiresIn', this.expiresIn.toISOString());
   }
 }
